Separate background from target models in PolarModel
