+++ include/EmpCyl2d.H
@@ -10,10 +10,13 @@
 
 #include <Eigen/Eigen>
 #include <Eigen/Eigenvalues>
-#include <gaussQ.H>
 
 #include <yaml-cpp/yaml.h>	// Parameters parsing
 #include <config_exp.h>		// For config macros
+
+#include <gaussQ.H>
+#include <EXPmath.H>
+#include <localmpi.H>
 
 // For reading and writing cache file
 //
@@ -69,6 +72,295 @@ public:
     double d_xi_to_r(double xi);
   };
 
+
+  class ExponCyl : public ModelCyl
+  {
+    
+  private:
+    
+    double acyl, sigma0;
+
+    // Assign values from YAML
+    //
+    void parse(const YAML::Node& conf)
+    {
+      try {
+	if (conf["acyl"]) 
+	  acyl = conf["acyl"].as<double>();
+	else
+	  acyl = 1.0;
+      }
+      catch (YAML::Exception & error) {
+	if (myid==0)
+	  std::cout << "Error parsing parameters in EmpCyl2d::ExponCyl: "
+		    << error.what() << std::endl
+		    << std::string(60, '-') << std::endl
+		    << "Config node"        << std::endl
+		    << std::string(60, '-') << std::endl
+		    << conf                 << std::endl
+		    << std::string(60, '-') << std::endl;
+	throw std::runtime_error("EmpCyl2d::ExponCyl: error parsing YAML config");
+      }
+    }
+    
+  public:
+
+    ExponCyl(const YAML::Node& par)
+    {
+      parse(par);
+      sigma0 = 0.5/(M_PI*acyl*acyl);
+      id = "expon";
+    }
+    
+    double pot(double r) {
+      double y = 0.5 * r / acyl;
+      return -M_PI*sigma0*r *
+	(EXPmath::cyl_bessel_i(0, y)*EXPmath::cyl_bessel_k(1, y) -
+	 EXPmath::cyl_bessel_i(1, y)*EXPmath::cyl_bessel_k(0, y));
+    }
+
+    double dpot(double r) {
+      double y = 0.5 * r / acyl;
+      return 2.0*M_PI*sigma0*y*
+	(EXPmath::cyl_bessel_i(0, y)*EXPmath::cyl_bessel_k(0, y) -
+	 EXPmath::cyl_bessel_i(1, y)*EXPmath::cyl_bessel_k(1, y));
+    }
+    
+    double dens(double r) {
+      return sigma0*exp(-r/acyl);
+    }
+    
+  };
+
+  class KuzminCyl : public ModelCyl
+  {
+  private:
+
+    double acyl;
+
+    // Assign values from YAML
+    //
+    void parse(const YAML::Node& conf)
+    {
+      try {
+	if (conf["acyl"]) 
+	  acyl = conf["acyl"].as<double>();
+	else
+	  acyl = 1.0;
+      }
+      catch (YAML::Exception & error) {
+	if (myid==0)
+	  std::cout << "Error parsing parameters in EmpCyl2d::KuzminCyl: "
+		    << error.what() << std::endl
+		    << std::string(60, '-') << std::endl
+		    << "Config node"        << std::endl
+		    << std::string(60, '-') << std::endl
+		    << conf                 << std::endl
+		    << std::string(60, '-') << std::endl;
+	throw std::runtime_error("EmpCyl2d::KuzminCyl: error parsing YAML config");
+      }
+    }
+    
+  public:
+  
+    KuzminCyl(const YAML::Node& par)
+    {
+      parse(par);
+      id = "kuzmin";
+    }
+
+    double pot(double R) {
+      double a2 = acyl * acyl;
+      return -1.0/sqrt(R*R + a2);
+    }
+    
+    double dpot(double R) {
+      double a2 = acyl * acyl;
+      return R/pow(R*R + a2, 1.5);
+    }
+    
+    double dens(double R) {
+      double a2 = acyl * acyl;
+      return 4.0*M_PI*acyl/pow(R*R + a2, 1.5)/(2.0*M_PI);
+      //     ^
+      //     |
+      // This 4pi from Poisson's eqn
+    }
+    
+  };
+
+
+  class MestelCyl : public ModelCyl
+  {
+  protected:
+
+    double vrot, rot;
+
+    // Assign values from YAML
+    //
+    void parse(const YAML::Node& conf)
+    {
+      try {
+	if (conf["vrot"]) 
+	  vrot = conf["vrot"].as<double>();
+	else
+	  vrot = 1.0;
+      }
+      catch (YAML::Exception & error) {
+	if (myid==0)
+	  std::cout << "Error parsing parameters in EmpCyl2d::MestelCyl: "
+		    << error.what() << std::endl
+		    << std::string(60, '-') << std::endl
+		    << "Config node"        << std::endl
+		    << std::string(60, '-') << std::endl
+		    << conf                 << std::endl
+		    << std::string(60, '-') << std::endl;
+	throw std::runtime_error("EmpCyl2d::MestelCyl: error parsing YAML config");
+      }
+    }
+    
+  public:
+  
+    MestelCyl(const YAML::Node& par)
+    {
+      parse(par);
+      rot = vrot*vrot;
+      id = "mestel";
+    }
+
+    virtual double pot(double R) {
+      if (R>0.0) return rot*log(R);
+      else throw std::runtime_error("MestelCyl::pot: R<=0");
+    }
+    
+    virtual double dpot(double R) {
+      if (R>0.0) return rot/R;
+      else throw std::runtime_error("MestelCyl::dpot: R<=0");
+    }
+
+    virtual double dens(double R) {
+      if (R>0.0) return rot/(2.0*M_PI*R);
+      else throw std::runtime_error("MestelCyl::dens: R<=0");
+    }
+  };
+  
+
+  class ZangCyl : public MestelCyl
+  {
+  
+  private:
+    //! Parameters
+    double mu, nu, ri, ro;
+
+    //! Softening factor (not currently used)
+    double asoft = 1.0e-8;
+
+    //! Ignore inner cut-off for N<0.05
+    bool Inner = true;
+
+    //! Taper factors
+    double Tifac, Tofac;
+
+    //! Inner taper function
+    double Tinner(double Jp)
+    {
+      double fac = pow(Jp, nu);
+      return fac/(Tifac + fac);
+    }
+
+    //! Outer taper function
+    double Touter(double Jp)
+    {
+      return 1.0/(1.0 + pow(Jp/Tofac, mu));
+    }
+    
+    //! Deriv of inner taper function
+    double dTinner(double Jp)
+    {
+      double fac  = pow(Jp, nu);
+      double fac2 = Tifac + fac;
+      return nu*fac/Jp/(fac2*fac2);
+    }
+    
+    //! Deriv of outer taper function
+    double dTouter(double Jp)
+    {
+      double fac = pow(Jp/Tofac, mu);
+      double fac2 = 1.0 + fac;
+      return -mu*fac/Jp/(fac2*fac2);
+    }
+    
+  protected:
+    
+    //! Assign values from YAML
+    void parse(const YAML::Node& conf)
+    {
+      try {
+	if (conf["Ninner"]) 
+	  nu = conf["Ninner"].as<double>();
+	else
+	  nu = 2.0;
+
+	if (conf["Mouter"]) 
+	  mu = conf["Mouter"].as<double>();
+	else
+	  mu = 2.0;
+	
+	if (conf["Ri"]) 
+	  ri = conf["Ri"].as<double>();
+	else
+	  ri = 1.0;
+	
+	if (conf["Ro"]) 
+	  ro = conf["Ro"].as<double>();
+	else
+	  ro = 10.0;
+      }
+      catch (YAML::Exception & error) {
+	if (myid==0)
+	  std::cout << "Error parsing parameters in EmpCyl2d::ZangCyl: "
+		    << error.what() << std::endl
+		    << std::string(60, '-') << std::endl
+		    << "Config node"        << std::endl
+		    << std::string(60, '-') << std::endl
+		    << conf                 << std::endl
+		    << std::string(60, '-') << std::endl;
+	throw std::runtime_error("EmpCyl2d::ZangCyl: error parsing YAML config");
+      }
+    }
+
+  public:
+  
+    //! Constructor
+    ZangCyl(const YAML::Node& par) : MestelCyl(par)
+    {
+      // Parse the YAML
+      parse(par);
+      
+      // Assign the id
+      id = "zang";
+      
+      // Cache taper factors
+      Tifac = pow(ri*vrot, nu);
+      Tofac = ro*vrot;
+      
+      if (nu<0.05) {
+	// Exponent is now for mapping only
+	Inner = false;
+      }
+    }
+    
+    //! Surface density
+    double dens(double R)
+    {
+      double ret = MestelCyl::dens(R) * Touter(R);
+      if (Inner) ret *= Tinner(R);
+      return ret;
+    }
+    
+  };
+
+
 protected:
 
   //! Contains parameter database
@@ -112,11 +404,6 @@ protected:
   std::shared_ptr<Basis2d>  basis;
 
   std::shared_ptr<ModelCyl> createModel();
-
-  class ExponCyl;
-  class MestelCyl;
-  class KuzminCyl;
-  class ZangCyl;
 
   //! Mapping instance
   Mapping map;
@@ -206,22 +493,15 @@ public:
   //! Check for configuration
   bool operator()() { return configured; }
 
-  //@{
-  //! Background evaluation
-  void background(double r, double & p, double & dr)
-  {
-    p  = disk->pot(r);
-    dr = disk->dpot(r);
-  }
-
-  std::tuple<double, double, double> background(double r)
-  {
-    return {disk->pot(r), disk->dpot(r), disk->dens(r)};
-  }
-  //@}
-
   //! Get coordinate mapping object
   Mapping getMapping() { return map; }
+
+  //! Get model name
+  const std::string getModelName() { return model; }
+
+  //! Get biorthogonal function base name
+  const std::string getBiorthName() { return biorth; }
+
 };
